Tweet-Application-master/src/app/authentication/users/users.component.ts
import { AuthenticationService } from './../authentication.service';
import { Component, OnInit } from '@angular/core';
import { User } from '../user.model';

@Component({
  selector: 'app-users',
  templateUrl: './users.component.html',
  styleUrls: ['./users.component.scss']
})
export class UsersComponent implements OnInit {

  users: User[];
  constructor(private authenticationService: AuthenticationService) { }

  ngOnInit(): void {
    this.authenticationService.getAllUser().subscribe(users=>{
      this.users = users;
    })
  }

}
